requirements.txt
fastapi
uvicorn
python-telegram-bot
openai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
apscheduler
pytz
python-dotenv
sqlalchemy
aiohttp
pydantic
python-multipart
